+++ src/App.tsx
@@ -3,7 +3,8 @@ import "./App.css";
 
 import { Banner } from "./components/Banner/Banner";
 import { CodeSyncProvider } from "./components/CodeSyncProvider/CodeSyncProvider";
-import { DownloadPdfWithTheme, LightThemeToggle } from "./components/LightThemeToggle/LightThemeToggle";
+import { DownloadPdfWithTheme } from "./components/DownloadThemedPdf/DownloadThemedPdf";
+import { LightThemeToggle } from "./components/LightThemeToggle/LightThemeToggle";
 import { type ILocationContext, LocationContext } from "./components/LocationProvider/LocationProvider";
 import { type IMetadataContext, MetadataContext } from "./components/MetadataProvider/MetadataProvider";
 import { NotesProvider } from "./components/NotesProvider/NotesProvider";
